streamlit_web_app/requirements.txt
matplotlib==3.6.2
pandas==2.2.2
python-dotenv==1.0.1
requests==2.28.1
SQLAlchemy==2.0.32
streamlit==1.37.1
numpy==1.25.2

